backend/requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
pytest
